mlstocks/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
contourpy==1.0.5
    # via matplotlib
cycler==0.11.0
    # via matplotlib
django==4.1.1
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
fonttools==4.37.3
    # via matplotlib
idna==3.4
    # via requests
joblib==1.2.0
    # via scikit-learn
kiwisolver==1.4.4
    # via matplotlib
lxml==4.9.1
    # via yfinance
matplotlib==3.6.0
    # via -r requirements.in
multitasking==0.0.11
    # via yfinance
mysqlclient==2.1.1
    # via -r requirements.in
numpy==1.23.3
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   yfinance
packaging==21.3
    # via matplotlib
pandas==1.5.0
    # via yfinance
pillow==9.2.0
    # via matplotlib
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2022.2.1
    # via
    #   djangorestframework
    #   pandas
requests==2.28.1
    # via yfinance
scikit-learn==1.1.2
    # via sklearn
scipy==1.9.1
    # via scikit-learn
six==1.16.0
    # via python-dateutil
sklearn==0.0
    # via -r requirements.in
sqlparse==0.4.2
    # via django
threadpoolctl==3.1.0
    # via scikit-learn
tzdata==2022.2
    # via django
urllib3==1.26.12
    # via requests
yfinance==0.1.74
    # via -r requirements.in
